refactor: created packages for ui and models in the filter package
