+++ app/src/main/java/com/example/househomey/filter/model/FilterCallback.java
@@ -1,4 +1,4 @@
-package com.example.househomey.Filters;
+package com.example.househomey.filter.model;
 
 
 /**
